Fix CPU and duration parsing in SDLBuilder; update default duration constraint
- Changed CPU units parsing from `parseInt` to `parseFloat` for better handling of decimal values.
- Updated duration parsing in SDLBuilder to use `parseFloat` instead of `parseInt`.
- Adjusted the default hours in `DURATION_CONSTRAINTS` from 24 to 1 for more accurate initial settings in deployment forms.
- Modified button alignment in Express and Streamlit calculator templates for improved UI consistency.

+++ app/app/services/backend/page.tsx
@@ -152,7 +152,7 @@ const SDLBuilder: React.FC = () => {
       const config: DeploymentConfig = {
         appPort: portNumber,
         deploymentDuration: `${values.deploymentDuration}h`,
-        appCpuUnits: parseInt(values.cpuValue),
+        appCpuUnits: parseFloat(values.cpuValue),
         appMemorySize: `${values.memoryValue}${values.memoryUnit}`,
         appStorageSize: `${values.ephemeralValue}${values.ephemeralUnit}`,
       };
@@ -802,7 +802,7 @@ const SDLBuilder: React.FC = () => {
                     setValues({ ...values, deploymentDuration: 0 });
                     setDurationError("Duration value is required");
                   } else {
-                    const value = parseInt(inputValue);
+                    const value = parseFloat(inputValue);
                     if (!isNaN(value)) {
                       setValues({ ...values, deploymentDuration: value });
                       validateDuration(value);
